+++ README.md
@@ -27,7 +27,7 @@ a jira-inspired to-do list
 
 ## Useful Links:
 * [User Guide](https://github.com/JRSmiffy/jara3/blob/main/docs/userguide.md)
-* [Technical References](https://github.com/JRSmiffy/jara3/blob/main/docs/technicalreferences.md)
+* [Technical References](https://github.com/JRSmiffy/jara3/blob/main/docs/techref.md)
 * Tools
     * [Miro](https://miro.com/app/board/uXjVOREpiSo=/)
     * [Jira](https://jrsmiffy.atlassian.net/jira/software/projects/J3/boards/1)
