+++ eureka-server/src/main/java/com/jrsmiffy/jara3/eurekaserver/EurekaServerApplication.java
@@ -1,0 +1,13 @@
+package com.jrsmiffy.jara3.eurekaserver;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class EurekaServerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(EurekaServerApplication.class, args);
+	}
+
+}
