+++ eureka-server/src/test/java/com/jrsmiffy/jara3/eurekaserver/EurekaServerApplicationTests.java
@@ -1,0 +1,13 @@
+package com.jrsmiffy.jara3.eurekaserver;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EurekaServerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
