--- jara3-eureka-server/src/main/java/com/jrsmiffy/jara3eurekaserver/Jara3EurekaServerApplication.java
@@ -1,13 +1,0 @@
-package com.jrsmiffy.jara3eurekaserver;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Jara3EurekaServerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Jara3EurekaServerApplication.class, args);
-	}
-
-}
