--- jara3-eureka-server/src/test/java/com/jrsmiffy/jara3eurekaserver/Jara3EurekaServerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.jrsmiffy.jara3eurekaserver;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Jara3EurekaServerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
