--- jara3-user-service/src/main/java/com/jrsmiffy/jara3userservice/Jara3UserServiceApplication.java
@@ -1,13 +1,0 @@
-package com.jrsmiffy.jara3userservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Jara3UserServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Jara3UserServiceApplication.class, args);
-	}
-
-}
