--- jara3-user-service/src/test/java/com/jrsmiffy/jara3userservice/Jara3UserServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.jrsmiffy.jara3userservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Jara3UserServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
