+++ user-service/src/main/java/com/jrsmiffy/jara3/userservice/UserServiceApplication.java
@@ -1,0 +1,13 @@
+package com.jrsmiffy.jara3.userservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class UserServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(UserServiceApplication.class, args);
+	}
+
+}
