+++ user-service/src/test/java/com/jrsmiffy/jara3/userservice/UserServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.jrsmiffy.jara3.userservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UserServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
