+++ web-app/src/favicon.ico
